Feature/calculator api

+++ src/main/java/com/tis/mx/application/dto/InitialInvestmentDto.java
@@ -17,7 +17,7 @@ package com.tis.mx.application.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-
+ 
 /**
  * The Class InitialInvestmentDto.
  */
@@ -26,10 +26,10 @@ import lombok.Setter;
 public class InitialInvestmentDto {
 
   /** The initial investment. */
-  private Double initialInvestment;
+  private Float initialInvestment;
   
   /** The yearly input. */
-  private Double yearlyInput;
+  private Float yearlyInput;
   
   /** The yearly input increment. */
   private Integer yearlyInputIncrement;
